Revert "Revert "Merge branch 'main' into Jonathan_UI""
This reverts commit 0d80cd6b3775b3970d575cd07239afbf5bf2e29b.

+++ Viking Attack/Assets/Scripts/Character_scripts/CharacterBase.cs
@@ -19,7 +19,7 @@ namespace ItemNamespace
         [SerializeField] private float attackCooldown;
         [SerializeField] private int damage;
         [SerializeField] private float chasingSpeedMultiplier;
-        [SerializeField] private int moveSpeed;
+        [SerializeField] private float moveSpeed;
         [SerializeField] private float maxHealth;
         [SerializeField] private float experienceRadius;
         [SerializeField] private float experience;
@@ -65,7 +65,7 @@ namespace ItemNamespace
             return chasingSpeedMultiplier;
         }
 
-        public int GetMovementSpeed()
+        public float GetMovementSpeed()
         {
             return moveSpeed;
         }
